pyproject.toml
[project]
name = "hand-painting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
]
